Move the code for path collections into a standalone object
For some reason the child mapping wasn't being cached when we used an
implicit class.  It's possible you can get this working with the implicit
classes, but an explicit class is probably less fragile and easier to
reason about.
